// model/create-risk-rules-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateRiskRulesData
 */
export interface CreateRiskRulesData {
    /**
     * Description of the rule
     * @type {string}
     * @memberof CreateRiskRulesData
     */
    'description': string;
    /**
     * Field to be used for the rule
     * @type {string}
     * @memberof CreateRiskRulesData
     */
    'field': string;
    /**
     * Value to be used for the rule
     * @type {string}
     * @memberof CreateRiskRulesData
     */
    'value': string;
}